backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pdfkit>=1.0.0",
    "pillow>=12.0.0",
    "pyjwt[crypto]>=2.10.1",
    "reportlab>=4.4.6",
]
